+++ CHANGELOG.md
@@ -2,11 +2,14 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased][unreleased]
-### Changed
-- Remove webpack to generate the server code and replace with babel to generate es5 code
+### Added
 - Add the `/schema` middleware to get the GraphQL schema
 - Add the GraphiQL web interface in the `ql` subdirectory of the server
 - Add gulp tasks to clean and to prepare the build directory
+
+### Changed
+- Remove webpack to generate the server code and replace with babel to generate es5 code
+- Rename gulp tasks to follow a naming convention "system:function"
 
 ## Initial import in master - 2015-10-04
 ### Added